build(rollup): simplify config
