+++ rollup.config.js
@@ -1,44 +1,102 @@
+import path from 'path'
+import alias from '@rollup/plugin-alias'
+import resolve from '@rollup/plugin-node-resolve'
+import commonjs from '@rollup/plugin-commonjs'
+import replace from '@rollup/plugin-replace'
+import jsonReader from '@rollup/plugin-json'
+import nodeBuiltins from 'rollup-plugin-node-builtins'
+import nodeGlobals from 'rollup-plugin-node-globals'
+import namedDirectory from 'rollup-plugin-named-directory'
+import html from 'rollup-plugin-html'
+import postcss from 'rollup-plugin-postcss'
+import babel from 'rollup-plugin-babel'
+import livereload from 'rollup-plugin-livereload'
+import serve from 'rollup-plugin-serve'
+import filesize from 'rollup-plugin-filesize'
+import size from 'rollup-plugin-sizes'
+import visualize from 'rollup-plugin-visualizer'
+import { terser } from 'rollup-plugin-terser'
+import pkg from './package.json'
 
-const analyze = require('rollup-plugin-analyzer')
-const { terser } = require('rollup-plugin-terser')
-
-const { baseConfig } = require('./rollup.config.base')
-
-const minimize = !!process.env.MINIMIZE
+const dev = !!process.env.DEV
 const analyzeBundle = !!process.env.ANALYZE_BUNDLE
+const minimize = !!process.env.MINIMIZE
 
-const output = [
-  ...baseConfig.output,
-  ...(minimize
-    ? [
-      {
-        file: 'dist/clappr-core.min.js',
-        format: 'umd',
-        exports: 'named',
-        name: 'Clappr',
-        sourcemap: true,
-        plugins: [
-          terser({
-            include: [/^.+\.min\.js$/],
-          }),
-        ],
-      },
-    ]
-    : []),
-  {
-    file: 'dist/clappr-core.esm.js',
-    format: 'esm',
-    exports: 'named',
-  },
-]
+const postcssOptions = {
+  use: [
+    ['sass', {
+      includePaths: [path.resolve('src/base/scss')]
+    }]
+  ]
+}
+const aliasPluginOptions = { entries: { 'clappr-zepto': 'node_modules/clappr-zepto/zepto.js', '@': __dirname + '/src' } }
+const replacePluginOptions = { VERSION: JSON.stringify(pkg.version) }
+const babelPluginOptions = { exclude: 'node_modules/**' }
+const servePluginOptions = { contentBase: ['dist', 'public'], host: '0.0.0.0', port: '8080' }
+const livereloadPluginOptions = { watch: ['dist', 'public'] }
+const visualizePluginOptions = { open: true }
+const terserPluginOptions = { include: [/^.+\.min\.js$/] }
 
 const plugins = [
-  ...baseConfig.plugins,
-  ...(analyzeBundle ? [analyze()] : []),
+  jsonReader(),
+  alias(aliasPluginOptions),
+  replace(replacePluginOptions),
+  resolve(),
+  commonjs(),
+  nodeBuiltins(),
+  nodeGlobals(),
+  babel(babelPluginOptions),
+  namedDirectory(),
+  html(),
+  postcss(postcssOptions),
+  size(),
+  filesize()
 ]
 
-module.exports = {
-  ...baseConfig,
-  output,
-  plugins,
-}
+dev && plugins.push(serve(servePluginOptions), livereload(livereloadPluginOptions))
+analyzeBundle && plugins.push(visualize(visualizePluginOptions))
+
+const rollupConfig = [
+  {
+    input: 'src/main.js',
+    output: {
+      exports: 'named',
+      name: 'Clappr',
+      file: pkg.main,
+      format: 'umd',
+      sourcemap: true,
+    },
+    plugins,
+  },
+  {
+    input: 'src/main.js',
+    output: {
+      exports: 'named',
+      name: 'Clappr',
+      file: pkg.module,
+      format: 'esm',
+    },
+    plugins,
+  }
+]
+
+minimize && rollupConfig.push(
+  {
+    input: 'src/main.js',
+    output: [
+      {
+        exports: 'named',
+        name: 'Clappr',
+        file: 'dist/clappr-core.min.js',
+        format: 'umd',
+        sourcemap: true,
+      },
+    ],
+    plugins: [
+      ...plugins,
+      terser(terserPluginOptions),
+    ],
+  }
+)
+
+export default rollupConfig
